requirements.txt
certifi==2021.10.8
cssselect2==0.4.1
da-vinci==0.3.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==2.2.5
django-admin-interface==0.18.5
django-autoslug==1.9.8
django-ckeditor==6.2.0
django-colorfield==0.6.3
django-crum==0.7.9
django-currentuser==0.5.3
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
django-solo==1.1.3
django-svg-image-form-field==1.0.1
easy-thumbnails==2.8.1
lxml==4.7.1
Pillow==9.0.0
#psycopg2==2.8.6
python-decouple==3.5
pytz==2021.3
reportlab==3.6.6
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
webencodings==0.5.1
wincertstore==0.2
